requirements.txt
# AI File Organiser - Python Dependencies
# Copyright © 2025 Alexandru Emanuel Vasile. All rights reserved.

# ==============================================================================
# DEPENDENCY INFORMATION
# ==============================================================================
# This project uses graceful degradation for optional packages.
# Core functionality works without optional dependencies, but some features
# will be disabled. The code includes proper import guards and type checking
# annotations following PEP 484 and PEP 561.
#
# REQUIRED packages: requests, PyYAML, fastapi, uvicorn, pydantic, aiofiles
# OPTIONAL packages: watchdog, python-docx, pydantic-settings, mutagen, Pillow
#
# To install all packages: pip install -r requirements.txt
# To install only required: pip install requests PyYAML fastapi uvicorn pydantic aiofiles
# ==============================================================================

# Core dependencies (REQUIRED)
requests>=2.31.0        # HTTP library for API calls (RFC 7230-7235)
PyYAML>=6.0.0          # YAML parser for configuration files
jsonschema>=4.19.0     # JSON Schema validation (draft 2020-12)

# Web dashboard (REQUIRED for UI)
fastapi>=0.104.0       # Modern web framework based on ASGI (PEP 3333)
uvicorn>=0.24.0        # ASGI server implementation
pydantic>=2.4.0        # Data validation using Python type hints (PEP 484)

# Async and caching (REQUIRED for performance)
aiofiles>=23.2.1       # Async file operations
diskcache>=5.6.0       # Persistent disk-based caching

# File monitoring (OPTIONAL - graceful degradation if missing)
# Reference: inotify (Linux), FSEvents (macOS), ReadDirectoryChangesW (Windows)
watchdog>=3.0.0        # Cross-platform filesystem event monitoring

# Configuration management (OPTIONAL)
# Reference: Pydantic Settings for configuration from env vars
pydantic-settings>=2.1.0  # Environment-based configuration
click>=8.1.0           # CLI framework

# Document processing (OPTIONAL - graceful degradation if missing)
# Reference: PyPDF2 for PDF manipulation, python-docx for DOCX (ISO/IEC 29500)
PyPDF2>=3.0.0          # PDF metadata and text extraction
python-docx>=1.0.0     # Microsoft Word DOCX file processing

# Advanced metadata extraction (OPTIONAL)
# Reference: EXIF (Exchangeable Image File Format), ID3 tags
Pillow>=10.0.0         # Image processing and EXIF data extraction
mutagen>=1.47.0        # Audio/video metadata (ID3, MP4, FLAC, etc.)

# Security (REQUIRED for license validation)
cryptography>=41.0.0   # Cryptographic primitives
bcrypt>=4.0.0          # Password hashing (Blowfish cipher)
PyJWT>=2.8.0           # JSON Web Tokens (RFC 7519)

# User interface (REQUIRED for progress display)
tqdm>=4.66.0           # Progress bars for terminal

# Testing (REQUIRED for development and CI/CD)
pytest>=7.4.0          # Testing framework
pytest-cov>=4.1.0      # Code coverage plugin
pytest-asyncio>=0.21.0 # Async testing support
pytest-mock>=3.11.1    # Mocking utilities

# Additional utilities (OPTIONAL)
organize>=0.1.0        # File organization rules engine
fs>=2.4.0              # PyFilesystem2 - unified filesystem API
filelock>=3.12.0       # Cross-platform file locking
watchfiles>=0.21.0     # Alternative file watcher (Rust-based, faster)
filetype>=1.2.0        # File type detection via magic numbers
python-magic>=0.4.27   # libmagic bindings for MIME detection

# ==============================================================================
# EXTERNAL DEPENDENCIES
# ==============================================================================
# Ollama (Large Language Model runtime)
# Required for AI-powered file classification
# Installation: Visit https://ollama.ai or run: python setup_ollama.py
# Reference: Ollama provides local LLM inference for classification
# ==============================================================================
